back to master

+++ ft_putnbr_hex_ptr.c
@@ -6,13 +6,13 @@
 /*   By: jschmitz <jschmitz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/27 13:30:23 by jschmitz          #+#    #+#             */
-/*   Updated: 2024/07/05 18:22:50 by jschmitz         ###   ########.fr       */
+/*   Updated: 2024/07/08 21:12:34 by jschmitz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-void	ft_putnbr_hex_ptr(unsigned long long nbr, int *count, char casetype)
+void	ft_putnbr_hex_ptr(unsigned long nbr, int *count, char casetype)
 {
 	char	*base;
